+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Year-by-year temperature chart. The old temperature chart is selectable through a drop-down
 * Selection of min, mean and max temperature through switches in year-by-year chart
+* Add pan and zoom in the x direction through mouse drag and mouse wheel to year-by-year temperature chart
 
 ### Changed
 * Rename page to Emil's Weather Station @ bergetvidhandfatet